+++ example/lib/providers/search_query_provider.pg.dart
@@ -24,9 +24,6 @@ class $SearchQueryFacadeRef {
 
   void invalidate() => _ref.invalidate(_provider);
 
-  void setState(String value) =>
-      _ref.read(_provider.notifier).updateState(value);
-
   void listen(
     void Function(String? previous, String next) listener, {
     void Function(Object, StackTrace)? onError,